etl-worker/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-financial-news-worker"
version = "1.0.0"
description = "A scalable RAG system for processing financial news using Kafka, Spark, and Chroma"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "RAG Worker Team", email = "team@rag-worker.com"}
]
maintainers = [
    {name = "RAG Worker Team", email = "team@rag-worker.com"}
]
keywords = ["rag", "financial", "news", "kafka", "spark", "chroma", "vector-database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: Linguistic",
]
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "pyspark>=4.0.0",
    "kafka-python>=2.0.2",
    "chromadb>=0.4.22",
    "pydantic>=2.5.2",
    "pyyaml>=6.0.1",

    # Data processing
    "pandas>=2.1.4",
    "numpy>=1.24.3",
    "requests>=2.31.0",

    # Vector embeddings
    "sentence-transformers>=2.2.2",
    "transformers>=4.36.2",
    "torch>=2.1.2",

    # LLM Integration
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "click>=8.1.7",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.3",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.1",

    # Development tools
    "black>=23.12.1",
    "isort>=5.13.2",
    "flake8>=6.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ruff>=0.1.6",

    # Documentation
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.8",
    "mkdocstrings[python]>=0.24.0",

    # Type stubs
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
]

test = [
    "pytest>=7.4.3",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.1",
]

docs = [
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.8",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
rag-worker = "src.main:main"

[project.urls]
Homepage = "https://github.com/your-org/rag-financial-news-worker"
Documentation = "https://rag-financial-news-worker.readthedocs.io"
Repository = "https://github.com/your-org/rag-financial-news-worker"
"Bug Tracker" = "https://github.com/your-org/rag-financial-news-worker/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/config",
    "/docker",
    "/tests",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

# UV configuration
[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.1",
    "black>=23.12.1",
    "isort>=5.13.2",
    "flake8>=6.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ruff>=0.1.6",
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.8",
    "mkdocstrings[python]>=0.24.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.12",
]

# UV configuration

# Black configuration
[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["src", "tests"]

# MyPy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "kafka.*",
    "chromadb.*",
    "pyspark.*",
    "sentence_transformers.*",
    "transformers.*",
    "torch.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
    "-v",
    "--tb=short",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

# Ruff configuration
[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

# Pre-commit configuration (commented out due to TOML parsing issues)
# [tool.pre-commit]
# repos = []

# MkDocs configuration
[tool.mkdocs]
site_name = "RAG Financial News Worker"
site_description = "A scalable RAG system for processing financial news"
site_author = "RAG Worker Team"
site_url = "https://rag-financial-news-worker.readthedocs.io"
repo_name = "your-org/rag-financial-news-worker"
repo_url = "https://github.com/your-org/rag-financial-news-worker"
edit_uri = "edit/main/docs/"
docs_dir = "docs"
site_dir = "site"

[tool.mkdocs.theme]
name = "material"

[tool.mkdocs.plugins."mkdocstrings".handlers.python.selection]
docstring_style = "google"
